Define the auditd_pkg such that it can be verified against
Currently thehre is a bug in which if process auditing is enabled,
auditd will be disabled. However, the variable in which auditd should be
defined is not present.

This commit adds packages to each OS Variables file, based on a cursory
review of the relevant package repos.

It has not been tested.