AD-HOC fix (ansible_managed): Allow multiline ansible_managed
Ansible allows denoting files as "ansible_managed" by including a
specific string in the file. In the implemented repository, this
appeared as follows:

```
ansible_managed =
  This file is managed by Ansible. Changes to it will not be persisted and should be made instead to the development
  node repository. See:

    https://github.com/path/to/repo
```

This allows a more substantial help for those looking to make a change
permanent, particularly when Ansible is not yet the canonical mechanism
for managing change across all systems in a fleet.

Currently there is a bug in these roles in which if the ansible_managed
string is multiline, it will not be commented out in the particular
files in which it's inserted. Instead, only the first line is commented
out.

This commit resolves that by using the comment filter supplied since
Ansible 2.0 to ensure this content is commented out, rather than simply
prefixing the line with '##' (or another related character).

== Design Notes ==

=== Use of specific decorator ===

This commit attempts to preserve the pattern of double #'ing comments by
supplying the decorator (comment prefix) of '## '.