AD-HOC feat (Configuration): Support expressing an OSQuery secret
When registering OSQuery with a central server the server requires
either a shared secret to be expressed, or mutual TLS authentication.
This commit implements one such mechanism for managing this identity
verification.

== Design Notes ==

=== File over Environment ===

It's possible to express the secret both as an environment variable,
and as a file. The most common mechanisms for supplying environment
variables the author has used are:

- /etc/environment
- systemd extension

However, neither suits in this case. The playbook will be run against
non-systemd hosts, and /etc/environment is inherited by all processes.

Accordingly, a file seemed the best mechanism to express the secret in a
compatible way. Additionally, it appears to be used in the kolide
tests against OSQuery anyway.